refactor: cleaned up main file but account querying needs fixing
